Revert "Add #ifdef declarations for NON_MATCHING code"
This reverts commit 42807160cfb33af9f3cf2fa6b0ba98f232208cdb.

It makes searching for NON_MATCHING comment descriptions much less
convenient in most text editors or IDEs, and we want the function
CSV to be the single source of truth for function statuses. Having
a function marked as matching but not built because of a stray #ifdef
would be bad.